Maps

+++ db/migrate/20260127155311_drop_days_table.rb
@@ -1,5 +1,5 @@
 class DropDaysTable < ActiveRecord::Migration[7.1]
   def change
-    drop_table :days
+    drop_table :days if table_exists?(:days)
   end
 end
